+++ client/sports-course/src/lib/config.ts
@@ -1,9 +1,8 @@
 const defaultApiBaseUrl =
-    process.env.NODE_ENV === "development"
+    import.meta.env.MODE === "development"
         ? "http://localhost:7001"
         : "https://api.omkooora.com";
 
 export const apiBaseUrl =
-    process.env.REACT_APP_API_URL ||
-    process.env.NEXT_PUBLIC_API_URL ||
+    import.meta.env.VITE_API_URL ||
     defaultApiBaseUrl;